Added installer for patented codecs
We build QtMoko image with patented codecs enabled, but we delete
the libraries from resulting tarball or jffs. This program embeds
the deleted libraries as resources and can restore them.

This way we dont distribute images with patented stuff and allow
quite simple installation of this stuff from GUI.

To build the package you have to do:
qbuild
qbuild packages

Simple qbuild does not trigger copying libraries for embeding
resources.